requirements.txt
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0
requests>=2.31.0
playwright>=1.40.0
Pillow>=10.0.0
pywin32>=311
pandas>=2.0.0
pyodbc>=4.0.0
python-dotenv>=1.0.0
